% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prizASP.R
\name{prizASP}
\alias{prizASP}
\title{Get the aspects of the the URL you provide}
\usage{
prizASP(TOKEN = NA, URL = NA)
}
\arguments{
\item{TOKEN}{Your API token}

\item{URL}{The URL of the page you want to tag for topics. Be sure to include http://}
}
\value{
a list, and on a successful API call, a list of lists
}
\description{
This function attempts to take a URL and use the API to tag it with topics
}

