% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prizRECENT.R
\name{prizRECENT}
\alias{prizRECENT}
\title{Get the recommended, relevant, recent content for a topic}
\usage{
prizRECENT(TOKEN = NA, TID = NA)
}
\arguments{
\item{TOKEN}{Your API token}

\item{TID}{The topic id, which you can look for using \code{prizURL()}}
}
\value{
dataframe
}
\description{
Get the recommended, relevant, recent content for a topic
}

