% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prizSIM.R
\name{prizSIM}
\alias{prizSIM}
\title{Get similar topics}
\usage{
prizSIM(TOKEN = NA, TID = NA)
}
\arguments{
\item{TOKEN}{character. Your API token}

\item{TID}{integer The id number of the topic you want to find simlarities to.}
}
\description{
This function attempts to take a URL and use the API to tag it with topics
}

