% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prizTID.R
\name{prizTID}
\alias{prizTID}
\title{Get the topics for a given URL}
\usage{
prizTID()
}
\arguments{
\item{TOKEN}{Your API token}

\item{URL}{The URL of the page you want to tag for topics. Be sure to include http://}
}
\value{
dataframe
}
\description{
This is a helper function to get the human list of topics and topic ids
}

